bare minimum of marder config
