--- addons/compat_rnt/comapt_rnt_marder/XEH_PREP.hpp
@@ -1,2 +1,0 @@
-PREP(closeRamp);
-PREP(openRamp);
