--- addons/compat_rnt/comapt_rnt_marder/functions/fnc_closeRamp.sqf
@@ -1,24 +1,0 @@
-#include "script_component.hpp"
-/**
- *  Author: Timi007
- *
- *  Description:
- *      Closes ramp of the Marder.
- *
- *  Parameter(s):
- *      0: OBJECT - Vehicle (Marder).
- *
- *  Returns:
- *      Nothing.
- *
- *  Example:
- *      this call ttt_compat_rnt_Marder_fnc_closeRamp
- *
- */
-
-params ["_veh"];
-
-_veh animateSource ["heck_luke_rotation", 0];
-_veh animate ["Hide_Knopf_Heck_luke", 1];
-
-//[_veh, 2] remoteExecCall ["lock", _veh];
