--- addons/compat_rnt/comapt_rnt_marder/script_component.hpp
@@ -1,3 +1,0 @@
-#define SUBCOMPONENT fuchs
-#define SUBCOMPONENT_BEAUTIFIED Fuchs
-#include "..\script_component.hpp"
