+++ addons/compat_rnt/compat_rnt_marder/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT marder
+#define SUBCOMPONENT_BEAUTIFIED Marder
+#include "..\script_component.hpp"
