+++ addons/compat_rnt/compat_rnt_marder/stringtable.xml
