+++ addons/compat_rnt/stringtable.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<Project name="TTT">
-    <Package name="compat_rnt">
-        <Key ID="STR_ttt_compat_rnt_CSTRING">
-            <English></English>
+<Project name="ttt">
+    <Package name="Compat_Rnt">
+        <Key ID="STR_ttt_Compat_Rnt">
         </Key>
     </Package>
 </Project>
